// api/src/models/common.ts
export enum Status {
  FAILED = "FAILED",
  SUCCESS = "SUCCESS",
}

export enum VerificationMethod {
  EMAIL = "EMAIL",
  WALLET_CONNECTOR = "WALLET_CONNECTOR",
}

export enum AssetMatchType {
  POLICY_ID = "POLICY_ID",
  POLICY_ASSET_ID = "POLICY_ASSET_ID",
}

export enum AddressType {
  PAYMENT = "PAYMENT",
  STAKE = "STAKE",
}